refactor: use routes dir as main files location
create separated directories for every path of router
add Messages comp as mediation

BREAKING CHANGE:
change import statement via index.js
rename main components related to the route  to index.js
move dependent components to appropriate subdirectories of routes dir
