+++ src/routes/Chats/components/index.js
@@ -1,0 +1,4 @@
+import Chat from "./Chat";
+import ChatList from "./ChatList";
+
+export { Chat, ChatList }
