--- src/routes/Home.js
@@ -1,3 +1,0 @@
-export function Home() {
-  return (<h1>Home</h1>);
-}
