+++ src/routes/Home/index.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Home() {
+  return (
+    <div>
+      Home
+    </div>
+  )
+}
+
+export default Home
