+++ src/routes/Messages/components/index.js
@@ -1,0 +1,5 @@
+import MessageList from './MessageList';
+import MessageForm from './MessageForm';
+import Message from './Message';
+
+export { MessageList, MessageForm, Message }
