--- src/routes/Profile/components/index.js
@@ -1,3 +1,0 @@
-import ThemeToggle from "./ThemeToggle";
-
-export default ThemeToggle
