+++ src/routes/index.js
@@ -1,0 +1,6 @@
+import Chats from "./Chats";
+import Messages from "./Messages";
+import Profile from "./Profile";
+import Home from "./Home";
+
+export {Chats, Messages, Profile, Home}
